requirements.txt
fire
mwt
pandas
pillow
python-telegram-bot
